mem-cache: Cast addr in compression bypass DPRINTF
Cast addr to (unsigned long long) in BaseCache DPRINTF to fix %#llx
format specifier mismatch on macOS Clang compilation.

+++ src/mem/cache/base.cc
@@ -1830,7 +1830,8 @@ BaseCache::allocateBlock(const PacketPtr pkt, PacketList &writebacks)
             DPRINTF(CacheComp,
                     "Bypassing compression on block fill for address %#llx "
                     "due to high MSHR occupancy (%d/%d)\n",
-                    addr, mshrQueue.occupancy(), mshrQueue.capacity());
+                    (unsigned long long)addr, mshrQueue.occupancy(),
+                    mshrQueue.capacity());
         } else {
             const auto comp_data =
                 compressor->compress(pkt->getConstPtr<uint64_t>(),
